Randomize available drugs
There is now a 1/5 chance that a given drug will not be available for sale or purchase wherever you go each day.

In playtesting news, I bought DMT for like $50 on day 1, bought all I could, and sold it on day 3 for about 7 grand total. That would be enough to pay off the shark, if I had a way in the game to do that, yet.