Give loan shark an office
You can now pay back the shark!

Or take out more loans, if you squander the first one.

It's a little goofy that he'll apply the same deadline for a new loan as the original loan, even if you pay off the first loan before asking for more, but I'm genuinely not sure if other versions of the game every bothered handling those things. Maybe later.